Config to mount 'tiddlers' volume.
There is no easy way to retain a persistent volume across EC2 instances on Elastic Beanstalk. It is trivial to create a new volume from a snapshot when Elastic Beanstalk instances are started using the 'BlockDeviceMapping' property in the configuration, but not to mount an existing volume.
Hand written commands must be used to find and mount the existing volume, though in this case I've copied and adjusted a code snippet from the internet.