Fixed skipping of credential check.
Empty strings have to be passed to the server command for the 'username' and 'password' arguments  in order to open up access to the next positional argument which is the 'host' argument.
The credentials check is meant to be skipped when empty strings are passed but actually the TiddlyWiki was prompting for a username and password that were both literally double quotes.
Unescaping the quotes stopped the credential check and produced the desired behaviour of skipping it.